requirement.txt
appdirs==1.4.3
asgiref==3.2.7
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.6
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
django-simpleui
# django-pagedown==2.0.3
# django-markdown-deux=1.0.5
django-redis
# django-crispy-forms=1.9.1
django-mptt=0.11.0
django-ckeditor=5.9.0
django-forms-bootstrap=3.1.0

{% comment %}暂时不添加全局搜索 django-haystack=2.8.1
whoosh=2.7.4 {% endcomment %}
autobahn-20.4.3 
channels-2.4.0 
daphne-2.5.0 
txaio-20.4.1

channels-redis-2.4.2